moved the manager_base in pymodaq_gui as we may have use of it within pymodaq_gui modules of some other customapp
