+++ packages/pymodaq_gui/src/pymodaq_gui/managers/manager_base.py
